Update yaml constructor and representer
